Move submit logic from SysTray into MainWindow, and make SettingsTab emit events when use_server is toggled
